feat:add one key modify restore
thanks @SONGjiemo

#511
