--- model/request/waf_onekeymod_del_req.go
@@ -1,5 +1,0 @@
-package request
-
-type WafOneKeyModDelReq struct {
-	Id string `json:"id"  form:"id"`
-}
