--- model/request/waf_onekeymod_detail.go
@@ -1,5 +1,0 @@
-package request
-
-type WafOneKeyModDetailReq struct {
-	Id string `json:"id"  form:"id"`
-}
