--- model/request/waf_onekeymod_do_req.go
@@ -1,5 +1,0 @@
-package request
-
-type WafDoOneKeyModReq struct {
-	FilePath string `json:"file_path"` //文件所在路径
-}
