--- model/request/waf_onekeymod_search.go
@@ -1,7 +1,0 @@
-package request
-
-import "SamWaf/model/common/request"
-
-type WafOneKeyModSearchReq struct {
-	request.PageInfo
-}
